requirements.txt
streamlit
PyYAML
pandas
bcrypt
streamlit_authenticator